requirements.txt
# Medical Education AI System - Dependencies
# Python 3.11+ required

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
PyMuPDF==1.23.21  # Fastest PDF extraction
pdfplumber==0.10.3  # Tables and complex layouts
pytesseract==0.3.10  # OCR for scanned PDFs
Pillow==10.2.0  # Image processing
pdf2image==1.16.3  # PDF to image conversion

# LLM & AI
langchain==0.1.4
langchain-community==0.0.16
langgraph==0.0.20
langsmith==0.0.87
sentence-transformers==2.3.1  # For PubMedBERT embeddings
transformers==4.37.2
torch==2.1.2
accelerate==0.26.1

# Vector Database
qdrant-client==1.7.3

# Knowledge Graph
neo4j==5.16.0
py2neo==2021.2.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
celery==5.3.6
python-redis-lock==4.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.26.0
aiohttp==3.9.1

# PubMed API
biopython==1.83

# Text Processing
nltk==3.8.1
spacy==3.7.2
unstructured==0.12.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For API testing

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
jupyter==1.0.0

# Monitoring
prometheus-client==0.19.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
click==8.1.7
